Refactor DataAvailabilityChecker to use repository methods
