+++ dataland-backend/src/test/kotlin/org/dataland/datalandbackend/services/DataAvailabilityCheckerTest.kt
@@ -1,7 +1,5 @@
 package org.dataland.datalandbackend.services
 
-import jakarta.persistence.EntityManager
-import jakarta.persistence.PersistenceContext
 import org.dataland.datalandbackend.DatalandBackend
 import org.dataland.datalandbackend.repositories.DataMetaInformationRepository
 import org.dataland.datalandbackend.repositories.DataPointMetaInformationRepository
@@ -55,9 +53,6 @@ class DataAvailabilityCheckerTest {
         }
     }
 
-    @PersistenceContext
-    private lateinit var entityManager: EntityManager
-
     @Autowired
     private lateinit var dataMetaInformationRepository: DataMetaInformationRepository
 
@@ -87,7 +82,12 @@ class DataAvailabilityCheckerTest {
 
     @BeforeEach
     fun setUp() {
-        dataAvailabilityChecker = DataAvailabilityChecker(entityManager, dataCompositionService, dataPointMetaInformationRepository)
+        dataAvailabilityChecker =
+            DataAvailabilityChecker(
+                dataCompositionService,
+                dataMetaInformationRepository,
+                dataPointMetaInformationRepository,
+            )
         whenever(specificationClient.listFrameworkSpecifications()).thenReturn(
             listOf(SimpleFrameworkSpecification(IdWithRef(framework, "dummy"), "Test Framework")),
         )
